// src/app/feature/credit-administration/model/personalLoan.ts
export class PersonalLoan {
    branchName: string;
    branchDistrict: string;
    branchMunVdc: string;
    branchWardNo: string;
    loanType: string;
    interestRate: string;
    serviceFee: string;
    tenureOfLoanInMonths: string;
    tenureOfLoanInYears: string;
    installmentAmount: string;
}